Changed the repositories accordingly
