+++ src/app.js
@@ -14,9 +14,8 @@ const importModels = require('./models');
 const createRoute = require('./routes/resources');
 const createAuthenticationRoute = require('./routes/authentication');
 const createLabReportsRoute = require('./routes/lab_reports');
-const createLabsRoute = require('./routes/labs');
 
-const createUserRepository = require('./repositories/users-repository');
+const createUserRepository = require('./repositories/users');
 const createAuthorizationVerifier = require('./middlewares/authorization-verifier');
 
 const { permissions } = require('./helpers/util');
@@ -26,21 +25,38 @@ module.exports = function getApp(sqlConnection, mongoConnection) {
 
   importModels(sqlConnection);
 
-  const adminsRepository = require('./repositories/admins-repository')(sqlConnection); // eslint-disable-line global-require
-  const teachersRepository = require('./repositories/teachers-repository')(sqlConnection); // eslint-disable-line global-require
-  const studentsRepository = require('./repositories/students-repository')(sqlConnection); // eslint-disable-line global-require
-  const coursesRepository = require('./repositories/courses-repository')(sqlConnection); // eslint-disable-line global-require
-  const groupsRepository = require('./repositories/groups-repository')(sqlConnection); // eslint-disable-line global-require
-  const specialtiesRepository = require('./repositories/specialties-repository')(sqlConnection); // eslint-disable-line global-require
+  const adminsRepository = require('./repositories/admins')(sqlConnection); // eslint-disable-line global-require
+  const teachersRepository = require('./repositories/teachers')(sqlConnection); // eslint-disable-line global-require
+  const studentsRepository = require('./repositories/students')(sqlConnection); // eslint-disable-line global-require
+  const coursesRepository = require('./repositories/courses')(sqlConnection); // eslint-disable-line global-require
+  const groupsRepository = require('./repositories/groups')(sqlConnection); // eslint-disable-line global-require
+  const specialtiesRepository = require('./repositories/specialties')(sqlConnection); // eslint-disable-line global-require
+
+  // eslint-disable-next-line global-require
+  const courseInstancesSpecialtiesRepository = require('./repositories/course-instances_specialties')(sqlConnection);
 
   const userRepository = createUserRepository(sqlConnection);
 
-  const labReportsRepository = require('./repositories/lab_reports-repository')(mongoConnection, sqlConnection); // eslint-disable-line global-require
-  const labTasksRepository = require('./repositories/labs-repository')(mongoConnection, sqlConnection); // eslint-disable-line global-require
+  const labReportsRepository = require('./repositories/lab_reports')(mongoConnection, sqlConnection); // eslint-disable-line global-require
+  const courseInstancesRepository = require('./repositories/course_instances')(mongoConnection, sqlConnection); // eslint-disable-line global-require
 
-  app.use(cors({
-    exposedHeaders: ['Content-Disposition', 'Content-Length', 'Content-MD5'],
-  }));
+  // eslint-disable-next-line global-require
+  const courseInstancesRoute = require('./routes/course_instances')(
+    courseInstancesRepository,
+    permissions('crud|ru|r'),
+  );
+
+  // eslint-disable-next-line global-require
+  const courseInstancesSpecialtiesRoute = require('./routes/course_instances-specialties')(
+    courseInstancesSpecialtiesRepository,
+    permissions('crud|r|r'),
+  );
+
+  app.use(
+    cors({
+      exposedHeaders: ['Content-Disposition', 'Content-Length', 'Content-MD5'],
+    }),
+  );
   app.use(bodyParser.json());
   app.use(bodyParser.urlencoded({ extended: true }));
 
@@ -59,7 +75,8 @@ module.exports = function getApp(sqlConnection, mongoConnection) {
   app.use('/api/v1/groups', createRoute(groupsRepository, permissions('crud|r|r')));
   app.use('/api/v1/specialties', createRoute(specialtiesRepository, permissions('crud|r|r')));
   app.use('/api/v1/lab_reports', createLabReportsRoute(labReportsRepository, permissions('crud|ru|cr|')));
-  app.use('/api/v1/labs', createLabsRoute(labTasksRepository, permissions('crud|ru|r|')));
+  app.use('/api/v1/course_instances', courseInstancesRoute);
+  app.use('/api/v1/course_instances-specialties', courseInstancesSpecialtiesRoute);
 
   app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
