+++ src/models/course_instance-specialty.js
@@ -4,16 +4,19 @@ module.exports = function defineCourseSpecialty(sequelize, DataTypes) {
   const CourseInstanceSpecialty = sequelize.define(
     'CourseInstanceSpecialty',
     {
+      id: {
+        type: DataTypes.INTEGER,
+        primaryKey: true,
+        allowNull: false,
+      },
       courseInstanceId: {
         type: DataTypes.INTEGER,
         field: 'course_instance_id',
-        primaryKey: true,
         allowNull: false,
       },
       specialtyId: {
         type: DataTypes.INTEGER,
         field: 'specialty_id',
-        primaryKey: true,
         allowNull: false,
       },
     },
@@ -23,8 +26,10 @@ module.exports = function defineCourseSpecialty(sequelize, DataTypes) {
     },
   );
 
-  // eslint-disable-next-line no-unused-vars
-  CourseInstanceSpecialty.associate = function associateCourseInstanceSpecialty(models) {};
+  CourseInstanceSpecialty.associate = function associateCourseInstanceSpecialty(models) {
+    CourseInstanceSpecialty.belongsTo(models.CourseInstance, { foreignKey: 'courseInstanceId' });
+    CourseInstanceSpecialty.belongsTo(models.Specialty, { foreignKey: 'specialtyId' });
+  };
 
   return CourseInstanceSpecialty;
 };
